Rename countryRegionData to types
